Move tests to their packages
